requirements.txt
newspaper3k==0.2.8
weaviate-client==3.3.0
rfc3339==6.2
